requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pyyaml>=6.0
mlflow>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pytest>=7.4.0
jupyter>=1.0.0
